// genpoe_latest/src/app/shared/models/poetransaction.ts
export class PoETransaction {
    // peers: string[];
    // chaincodeName: string = Chaincodes.CHAINCODE_PMPOE;
    fileName: string;
    fileType: string;
    fileSize: string;
    documentType: string;
    sha256Hash: string;
    base64: string;
    sha1Hash: string;
    issuedTo: string;
    issuedByUser: string;
    issuedByOrg: string;
    allowStorage: string;
}