// genpoe_latest/src/app/shared/models/poetransactionorg.ts
export class PoETransactionOrg {
    // peers: string[];
    // chaincodeName: string = Chaincodes.CHAINCODE_PMPOE;
    fileName: string;
    fileType: string;
    fileSize: string;
    documentType: string;
    sha256Hash: string;
    base64: string;
    sha1Hash: string;
    issuedTo: string;
    issuedByUser: string;
    issuedByOrg: string;
    allowStorage: string;



    uniqueID: string;
    recipientEmail: string;
    recipientName: string;
​    recipientMobile: string;
}